+++ app/src/main/java/com/poli/prevendasomie/data/remote/responses/pedidos/IcmsSn.kt
@@ -4,15 +4,15 @@ import com.google.gson.annotations.SerializedName
 
 data class IcmsSn(
     @SerializedName("aliq_icms_sn")
-    val aliqIcmsSn: Int?,
+    val aliqIcmsSn: Double?,
     @SerializedName("base_icms_sn")
-    val baseIcmsSn: Int?,
+    val baseIcmsSn: Double?,
     @SerializedName("cod_sit_trib_icms_sn")
     val codSitTribIcmsSn: String?,
     @SerializedName("origem_icms_sn")
     val origemIcmsSn: String?,
     @SerializedName("valor_credito_icms_sn")
-    val valorCreditoIcmsSn: Int?,
+    val valorCreditoIcmsSn: Double?,
     @SerializedName("valor_icms_sn")
-    val valorIcmsSn: Int?
+    val valorIcmsSn: Double?
 )
