+++ app/src/main/java/com/poli/prevendasomie/data/remote/responses/pedidos/PisPadrao.kt
@@ -4,9 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 data class PisPadrao(
     @SerializedName("aliq_pis")
-    val aliqPis: Int?,
+    val aliqPis: Double?,
     @SerializedName("base_pis")
-    val basePis: Int?,
+    val basePis: Double?,
     @SerializedName("cod_sit_trib_pis")
     val codSitTribPis: String?,
     @SerializedName("qtde_unid_trib_pis")
@@ -14,7 +14,7 @@ data class PisPadrao(
     @SerializedName("tipo_calculo_pis")
     val tipoCalculoPis: String?,
     @SerializedName("valor_pis")
-    val valorPis: Int?,
+    val valorPis: Double?,
     @SerializedName("valor_unid_trib_pis")
-    val valorUnidTribPis: Int?
+    val valorUnidTribPis: Double?
 )
